_legacy_v1/api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.6.0
    # via
    #   django
    #   djoser
azure-core==1.25.0
    # via
    #   azure-storage-blob
    #   msrest
azure-storage-blob==12.13.1
    # via django-storages
blessed==1.20.0
    # via -r requirements.in
boto3==1.28.3
    # via -r requirements.in
botocore==1.31.3
    # via
    #   boto3
    #   s3transfer
certifi==2022.6.15
    # via
    #   msrest
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.0
    # via requests
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
croniter==1.3.15
    # via -r requirements.in
cryptography==37.0.4
    # via
    #   azure-storage-blob
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==4.2.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-guardian
    #   django-picklefield
    #   django-q2
    #   django-sendgrid-v5
    #   django-storages
    #   djangorestframework
    #   djangorestframework-guardian2
    #   djangorestframework-simplejwt
    #   drf-nested-routers
django-cors-headers==4.2.0
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
django-guardian==2.4.0
    # via
    #   -r requirements.in
    #   djangorestframework-guardian2
django-picklefield==3.1
    # via django-q2
django-q2==1.5.5
    # via -r requirements.in
django-rest-framework-tricks==0.2.14
    # via -r requirements.in
django-sendgrid-v5==1.2.2
    # via -r requirements.in
django-storages[azure]==1.13.2
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   django-rest-framework-tricks
    #   djangorestframework-guardian2
    #   djangorestframework-simplejwt
    #   drf-nested-routers
djangorestframework-guardian2==0.5.0
    # via -r requirements.in
djangorestframework-simplejwt==4.8.0
    # via djoser
djoser==2.1.0
    # via -r requirements.in
drf-nested-routers==0.93.4
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
faker==19.1.0
    # via -r requirements.in
gevent==23.7.0
    # via -r requirements.in
greenlet==2.0.2
    # via gevent
gunicorn==20.1.0
    # via -r requirements.in
humanize==4.6.0
    # via django-rest-framework-tricks
idna==3.3
    # via requests
inflection==0.5.1
    # via -r requirements.in
isodate==0.6.1
    # via msrest
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mailchimp3==3.0.21
    # via -r requirements.in
markupsafe==2.1.3
    # via jinja2
msrest==0.7.1
    # via azure-storage-blob
oauthlib==3.2.0
    # via
    #   requests-oauthlib
    #   social-auth-core
openpyxl==3.1.2
    # via -r requirements.in
pillow==10.0.1
    # via -r requirements.in
psutil==5.9.5
    # via -r requirements.in
psycopg2-binary==2.9.6
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.4.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python-dateutil==2.8.2
    # via
    #   botocore
    #   croniter
    #   faker
python-http-client==3.3.7
    # via
    #   django-sendgrid-v5
    #   sendgrid
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.1
    # via djangorestframework
pyyaml==6.0
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   azure-core
    #   coreapi
    #   mailchimp3
    #   msrest
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   msrest
    #   social-auth-core
s3transfer==0.6.0
    # via boto3
sendgrid==6.9.7
    # via django-sendgrid-v5
sentry-sdk==1.28.1
    # via -r requirements.in
six==1.16.0
    # via
    #   azure-core
    #   blessed
    #   isodate
    #   python-dateutil
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.3.0
    # via social-auth-app-django
sqlparse==0.4.2
    # via django
starkbank-ecdsa==2.0.3
    # via sendgrid
typing-extensions==4.3.0
    # via azure-core
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.11
    # via
    #   botocore
    #   requests
    #   sentry-sdk
wcwidth==0.2.5
    # via blessed
whitenoise==6.5.0
    # via -r requirements.in
zope-event==4.5.0
    # via gevent
zope-interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
